ARM: dts: msm: Defer regulator disable for msm8953 GPU SMMU
IOMMU driver takes vote on regulator for every unmap, attach or
detach call and removes it once done.

This is an overhead if GPU is in power collapse mode and huge
number of unmap calls are pending from a given context. It would
end up in turning ON/OFF the regulator those many times.

To optimize this, do a deferred regulator disable about 80ms so that
all subsequent requests to regulator enable will be just refcounted

Change-Id: I69d5daf962cd1926b1e7d2717947afdf2280542b
Signed-off-by: Sunil Khatri <sunilkh@codeaurora.org>